+++ lib/smtp_server/gmail.dart
@@ -1,4 +1,7 @@
 import '../smtp_server.dart';
 
 SmtpServer gmail(String username, String password) =>
-    new SmtpServer('smtp.gmail.com', username: username, password: password);
+    SmtpServer('smtp.gmail.com', username: username, password: password);
+
+SmtpServer gmailXoauth2(String token) =>
+    SmtpServer('smtp.gmail.com', xoauth2Token: token);