+++ lib/smtp_server/yahoo.dart
@@ -1,5 +1,5 @@
 import '../smtp_server.dart';
 
 SmtpServer yahoo(String username, String password) =>
-    new SmtpServer('smtp.mail.yahoo.com',
+    SmtpServer('smtp.mail.yahoo.com',
         port: 465, username: username, password: password);