+++ lib/src/entities/send_report.dart
@@ -1,10 +1,18 @@
 import 'message.dart';
-import 'problem.dart';
 
 class SendReport {
   final Message mail;
-  final bool sent;
-  final List<Problem> validationProblems;
+  final DateTime connectionOpened;
+  final DateTime messageSendingStart;
+  final DateTime messageSendingEnd;
 
-  SendReport(this.mail, this.sent, {this.validationProblems});
+  SendReport(this.mail, this.connectionOpened, this.messageSendingStart,
+      this.messageSendingEnd);
+
+  @override
+  String toString() {
+    return 'Message successfully sent.\n'
+        'Connection was opened at: $connectionOpened.\n'
+        'Sending the message started at: $messageSendingStart and finished at: $messageSendingEnd.';
+  }
 }