+++ lib/src/smtp/internal_representation/conversion.dart
@@ -1,6 +1,5 @@
 import 'dart:async';
-
-import 'package:dart2_constant/convert.dart' as convert;
+import 'dart:convert' as convert;
 
 const String eol = '\r\n';
 
@@ -8,7 +7,7 @@ List<int> to8(String s) => convert.utf8.encode(s);
 final List<int> eol8 = to8(eol);
 
 bool _isMultiByteContinuationByte(int b) {
-  // MultiByte continuation bytes start are 0b10XXXXXX.
+  // MultiByte continuation bytes are 0b10XXXXXX.
   return b >= 128 && b < 192;
 }
 
@@ -17,12 +16,29 @@ bool _isMultiByteContinuationByte(int b) {
 /// If [maxLength] is provided returns chunks where the maximum length is
 /// at most [maxLength].  (Overrides constructor argument).
 ///
-/// If [maxLength] < 4 and there is a longer multibyte character the returned
-/// chunk will be the complete multibyte and therefore possibly too long.
+/// If [maxLength] < 4 and there is a longer multiByte character the returned
+/// chunk will be the complete multiByte and therefore possibly too long.
+//
+// UTF8 multi-byte rules (copied from:
+// https://stackoverflow.com/questions/9356169/utf-8-continuation-bytes)
+//
+// The basic rules are this:
+//    If a byte starts with a 0 bit, it's a single byte value less than 128.
+//    If it starts with 11, it's the first byte of a multi-byte sequence and the
+//      number of 1 bits at the start indicates how many bytes there are in
+//      total (110xxxxx has two bytes, 1110xxxx has three and 11110xxx has four).
+//    If it starts with 10, it's a continuation byte.
+//
+// A 4 byte multi byte character for example is:
+// 11110000 10010000 10001101 10001000
+// Note that the 2nd, 3rd and 4th byte all start with 10
 Iterable<List<int>> split(List<int> data, int maxLength,
     {bool avoidUtf8Cut = true}) sync* {
   int start = 0;
   for (;;) {
+    if (start >= data.length) break;
+
+    // When using List.sublist "end" is not included in the output.
     int end = start + maxLength;
     if (end >= data.length) {
       yield data.sublist(start);
@@ -31,16 +47,20 @@ Iterable<List<int>> split(List<int> data, int maxLength,
 
     // Look at the character immediately following the chunk we would like to
     // return.
-    int e = end + 1;
+    int e = end; // We do not need to add 1!  e now "points" to the character
+    // behind the characters a List.subList would return.
+    // We have already verified that this character exists.
     while (avoidUtf8Cut && e > start && _isMultiByteContinuationByte(data[e])) {
       e--;
     }
 
     if (e == start) {
-      // We must return the complete multibyte, even though it's longer than
+      // We must return the complete multiByte, even though it's longer than
       // the requested size.
       e = start + 1;
-      while (_isMultiByteContinuationByte(data[e])) e++;
+      while (e < data.length && _isMultiByteContinuationByte(data[e])) {
+        e++;
+      }
     }
 
     yield data.sublist(start, e);
@@ -52,7 +72,7 @@ Stream<List<int>> _splitS(
     Stream<List<int>> dataS, int splitOver, int maxLength) {
   int currentLineLength = 0;
 
-  var sc = new StreamController<List<int>>();
+  var sc = StreamController<List<int>>();
   void processData(List<int> data) {
     if (data.length + currentLineLength > maxLength) {
       int targetLength = maxLength ~/ 2;
@@ -76,10 +96,7 @@ Stream<List<int>> _splitS(
   return sc.stream;
 }
 
-// Replace `_StreamTransformerBase` with `StreamTransformerBase`
-// when dart1 is no longer supported.
-// Also remove the _* helper classes.
-class StreamSplitter extends _StreamTransformerBase<List<int>, List<int>> {
+class StreamSplitter extends StreamTransformerBase<List<int>, List<int>> {
   final int maxLength;
   final int splitOverLength;
 
@@ -88,11 +105,4 @@ class StreamSplitter extends _StreamTransformerBase<List<int>, List<int>> {
   @override
   Stream<List<int>> bind(Stream<List<int>> stream) =>
       _splitS(stream, splitOverLength, maxLength);
-}
-
-abstract class _StreamTransformerBase<SS, ST>
-    implements StreamTransformer<SS, ST> {
-  @override
-  StreamTransformer<RS, RT> cast<RS, RT>() => throw new UnimplementedError(
-      'This dart2 functionality is not implemented');
 }