+++ lib/src/smtp/internal_representation/internal_representation.dart
@@ -1,15 +1,14 @@
 import 'dart:async';
-import 'dart:convert' as dart_convert;
+import 'dart:convert' as convert;
 
-import 'package:dart2_constant/convert.dart' as convert;
 import 'package:intl/intl.dart';
+import 'package:mailer/src/utils.dart';
 
-import 'conversion.dart';
-import '../capabilities.dart';
 import '../../entities/address.dart';
 import '../../entities/attachment.dart';
 import '../../entities/message.dart';
-import '../../entities/problem.dart';
+import '../capabilities.dart';
+import 'conversion.dart';
 
 part 'ir_content.dart';
 part 'ir_header.dart';
@@ -33,12 +32,6 @@ class _IRMetaInformation {
   final Capabilities capabilities;
 
   _IRMetaInformation(this.capabilities);
-}
-
-class IRProblemException implements Exception {
-  final Problem problem;
-
-  IRProblemException(this.problem);
 }
 
 abstract class _IROutput {