+++ lib/src/smtp/internal_representation/ir_header.dart
@@ -7,7 +7,7 @@ abstract class _IRHeader extends _IROutput {
   static final List<int> _b64postfix = convert.utf8.encode('?=$eol');
   static final int _b64Length = _b64prefix.length + _b64postfix.length;
 
-  Stream<List<int>> _outValue(String value) => new Stream.fromIterable(
+  Stream<List<int>> _outValue(String value) => Stream.fromIterable(
       [_name, ': ', value ?? '', eol].map(convert.utf8.encode));
 
   // Outputs value encoded as base64.
@@ -45,13 +45,6 @@ abstract class _IRHeader extends _IROutput {
 class _IRHeaderText extends _IRHeader {
   String _value;
 
-  // RegExp created using:
-  // https://mothereff.in/regexpu
-  // `string.match(/^[^\p{Cc}\p{Cf}\p{Zl}\p{Zp}]*$/u);`
-  // as suggested by: https://stackoverflow.com/questions/12052825/regular-expression-for-all-printable-characters-in-javascript
-  final _isPrintableRegExp = new RegExp(
-      r'^(?:[ -~\xA0-\xAC\xAE-\u05FF\u0606-\u061B\u061D-\u06DC\u06DE-\u070E\u0710-\u08E1\u08E3-\u180D\u180F-\u200A\u2010-\u2027\u202F-\u205F\u2065\u2070-\uD7FF\uE000-\uFEFE\uFF00-\uFFF8\uFFFC-\uFFFF]|[\uD800-\uD803\uD805-\uD82E\uD830-\uD833\uD835-\uDB3F\uDB41-\uDBFF][\uDC00-\uDFFF]|\uD804[\uDC00-\uDCBC\uDCBE-\uDCCC\uDCCE-\uDFFF]|\uD82F[\uDC00-\uDC9F\uDCA4-\uDFFF]|\uD834[\uDC00-\uDD72\uDD7B-\uDFFF]|\uDB40[\uDC00\uDC02-\uDC1F\uDC80-\uDFFF]|[\uD800-\uDBFF](?![\uDC00-\uDFFF])|(?:[^\uD800-\uDBFF]|^)[\uDC00-\uDFFF])*$');
-
   _IRHeaderText(String name, this._value) : super(name);
 
   @override
@@ -59,10 +52,10 @@ class _IRHeaderText extends _IRHeader {
     bool utf8Allowed = irMetaInformation.capabilities.smtpUtf8;
 
     if ((_value?.length ?? 0) > maxLineLength ||
-        !_isPrintableRegExp.hasMatch(_value) ||
+        !isPrintableRegExp.hasMatch(_value) ||
         // Make sure that text which looks like an encoded text is encoded.
         _value.contains('=?') ||
-        (!utf8Allowed && _value.contains(new RegExp(r'[^\x20-\x7E]')))) {
+        (!utf8Allowed && _value.contains(RegExp(r'[^\x20-\x7E]')))) {
       return _outValueB64(_value);
     }
     return _outValue(_value);
@@ -116,7 +109,7 @@ class _IRHeaderDate extends _IRHeader {
   final DateTime _dateTime;
 
   static final DateFormat _dateFormat =
-      new DateFormat('EEE, dd MMM yyyy HH:mm:ss +0000');
+      DateFormat('EEE, dd MMM yyyy HH:mm:ss +0000');
 
   _IRHeaderDate(String name, this._dateTime) : super(name);
 
@@ -137,48 +130,50 @@ Iterable<_IRHeader> _buildHeaders(Message message) {
     if (noCustom.contains(name)) return;
 
     if (value is String && value.contains('@')) {
-      headers.add(new _IRHeaderAddress(name, new Address(value)));
+      headers.add(_IRHeaderAddress(name, Address(value)));
     } else if (value is String) {
-      headers.add(new _IRHeaderText(name, value));
+      headers.add(_IRHeaderText(name, value));
     } else if (value is DateTime) {
-      headers.add(new _IRHeaderDate(name, value));
+      headers.add(_IRHeaderDate(name, value));
     } else if (value is Address) {
-      headers.add(new _IRHeaderAddress(name, value));
+      headers.add(_IRHeaderAddress(name, value));
     } else if (value is Iterable<Address>) {
-      headers.add(new _IRHeaderAddresses(name, value));
+      headers.add(_IRHeaderAddresses(name, value));
     } else if (value is Iterable<String> &&
         value.every((s) => (s ?? '').contains('@'))) {
-      headers
-          .add(new _IRHeaderAddresses(name, value.map((a) => new Address(a))));
+      headers.add(_IRHeaderAddresses(name, value.map((a) => Address(a))));
     } else {
-      throw new IRProblemException(
-          new Problem('INVALID_HEADER', 'Type of value for $name is invalid'));
+      throw InvalidHeaderException('Type of value for $name is invalid');
     }
   });
 
-  if (!msgHeader.containsKey('subject') && message.subject != null)
-    headers.add(new _IRHeaderText('subject', message.subject));
+  if (!msgHeader.containsKey('subject') && message.subject != null) {
+    headers.add(_IRHeaderText('subject', message.subject));
+  }
 
-  if (!msgHeader.containsKey('from'))
-    headers.add(new _IRHeaderAddress('from', message.fromAsAddress));
+  if (!msgHeader.containsKey('from')) {
+    headers.add(_IRHeaderAddress('from', message.fromAsAddress));
+  }
 
   if (!msgHeader.containsKey('to')) {
     var tos = message.recipientsAsAddresses ?? [];
-    if (tos.isNotEmpty) headers.add(new _IRHeaderAddresses('to', tos));
+    if (tos.isNotEmpty) headers.add(_IRHeaderAddresses('to', tos));
   }
 
   if (!msgHeader.containsKey('cc')) {
     var ccs = message.ccsAsAddresses ?? [];
-    if (ccs.isNotEmpty) headers.add(new _IRHeaderAddresses('cc', ccs));
+    if (ccs.isNotEmpty) headers.add(_IRHeaderAddresses('cc', ccs));
   }
 
-  if (!msgHeader.containsKey('date'))
-    headers.add(new _IRHeaderDate('date', new DateTime.now()));
+  if (!msgHeader.containsKey('date')) {
+    headers.add(_IRHeaderDate('date', DateTime.now()));
+  }
 
-  if (!msgHeader.containsKey('x-mailer'))
-    headers.add(new _IRHeaderText('x-mailer', 'Dart Mailer library 2'));
+  if (!msgHeader.containsKey('x-mailer')) {
+    headers.add(_IRHeaderText('x-mailer', 'Dart Mailer library 2'));
+  }
 
-  headers.add(new _IRHeaderText('mime-version', '1.0'));
+  headers.add(_IRHeaderText('mime-version', '1.0'));
 
   return headers;
 }