+++ lib/src/smtp/server_response.dart
@@ -1,8 +1,0 @@
-class ServerResponse {
-  final String responseCode;
-
-  /// Every line received from the server is one entry in the message list.
-  final List<String> responseLines;
-
-  ServerResponse(this.responseCode, this.responseLines);
-}