+++ lib/src/smtp/smtp_client.dart
@@ -1,175 +1,189 @@
 import 'dart:async';
-import 'package:dart2_constant/convert.dart' as convert;
-import 'package:logging/logging.dart';
+import 'dart:convert' as convert;
+import 'dart:io';
+
 import 'package:mailer/smtp_server.dart';
-import 'internal_representation/internal_representation.dart';
+
 import '../entities/message.dart';
-import '../entities/send_report.dart';
-import 'connection.dart';
 import 'capabilities.dart';
+import 'connection.dart';
 import 'exceptions.dart';
-import 'package:mailer/src/entities/problem.dart';
-import 'validator.dart';
+import 'internal_representation/internal_representation.dart';
 
-final Logger _logger = new Logger('smtp-client');
+/// Returns if ehlo was successful.
+Future<bool> _doEhlo(Connection c, String clientName) async {
+  var respEhlo = await c.send('EHLO $clientName', acceptedRespCodes: null);
 
-class SmtpClient {
-  final Connection _c;
-  final SmtpServer _smtpServer;
-
-  SmtpClient(this._smtpServer) : _c = new Connection(_smtpServer);
-
-  /// Returns the capabilities of the server if ehlo was successful.  null if
-  /// `helo` is necessary.
-  Future<Capabilities> _doEhlo() async {
-    var respEhlo =
-        await _c.send('EHLO ${_smtpServer.name}', acceptedRespCodes: null);
-
-    if (!respEhlo.responseCode.startsWith('2')) {
-      return null;
-    }
-
-    var capabilities = new Capabilities.fromResponse(respEhlo.responseLines);
-
-    if (!capabilities.startTls || _c.isSecure) {
-      return capabilities;
-    }
-
-    // Use a secure socket.  Server announced STARTTLS.
-    // The server supports TLS and we haven't switched to it yet,
-    // so let's do it.
-    var tlsResp = await _c.send('STARTTLS', acceptedRespCodes: null);
-    if (!tlsResp.responseCode.startsWith('2')) {
-      // Even though server announced STARTTLS, it now chickens out.
-      return null;
-    }
-
-    // Replace _socket with an encrypted version.
-    await _c.upgradeConnection();
-
-    // Restart EHLO process.  This time on a secure connection.
-    return _doEhlo();
+  if (!respEhlo.responseCode.startsWith('2')) {
+    return false;
   }
 
-  Future<Capabilities> _doEhloHelo() async {
-    var ehlo = await _doEhlo();
+  var capabilities = new Capabilities.fromResponse(respEhlo.responseLines);
 
-    if (ehlo != null) {
-      return ehlo;
-    }
-
-    // EHLO not accepted.  Let's try HELO.
-    await _c.send('HELO ${_smtpServer.name}');
-    return new Capabilities();
+  if (!capabilities.startTls || c.isSecure) {
+    c.capabilities = capabilities;
+    return true;
   }
 
-  Future<Null> _doAuthentication(Capabilities capabilities) async {
-    if (_smtpServer.username == null) {
-      return;
-    }
-
-    if (!capabilities.authLogin) {
-      throw new SmtpClientCommunicationException(
-          'The server does not support LOGIN authentication method.');
-    }
-
-    var username = _smtpServer.username;
-    var password = _smtpServer.password;
-
-    // 'Username:' in base64 is: VXN...
-    await _c.send('AUTH LOGIN',
-        acceptedRespCodes: ['334'], expect: 'VXNlcm5hbWU6');
-    // 'Password:' in base64 is: UGF...
-    await _c.send(convert.base64.encode(username.codeUnits),
-        acceptedRespCodes: ['334'], expect: 'UGFzc3dvcmQ6');
-    var loginResp = await _c
-        .send(convert.base64.encode(password.codeUnits), acceptedRespCodes: []);
-    if (!loginResp.responseCode.startsWith('2')) {
-      throw new SmtpClientAuthenticationException(
-          'Incorrect username ($username) / password');
-    }
+  // Use a secure socket.  Server announced STARTTLS.
+  // The server supports TLS and we haven't switched to it yet,
+  // so let's do it.
+  var tlsResp = await c.send('STARTTLS', acceptedRespCodes: null);
+  if (!tlsResp.responseCode.startsWith('2')) {
+    // Even though server announced STARTTLS, it now chickens out.
+    return false;
   }
 
-  Future<List<SendReport>> send(Message message) async {
-    final List<SendReport> sendReports = [];
+  // Replace _socket with an encrypted version.
+  await c.upgradeConnection();
 
-    // Don't even try to connect to the server, if message-validation fails.
-    var problems = validate(message);
-    if (problems.isNotEmpty) {
-      sendReports
-          .add(new SendReport(message, false, validationProblems: problems));
-      return sendReports;
-    }
+  // Restart EHLO process.  This time on a secure connection.
+  return _doEhlo(c, clientName);
+}
 
-    IRMessage irMessage;
+Future<void> _doEhloHelo(Connection c, {String clientName}) async {
+  if (clientName == null || clientName.trim().isEmpty) {
+    clientName = Platform.localHostname;
+  }
+
+  final ehloSuccessful = await _doEhlo(c, clientName);
+
+  if (ehloSuccessful) {
+    return;
+  }
+
+  // EHLO not accepted.  Let's try HELO.
+  await c.send('HELO $clientName');
+  c.capabilities = new Capabilities();
+}
+
+Future<bool> _doAuthLogin(Connection c) async {
+  var capabilities = c.capabilities;
+  if (!capabilities.authLogin) {
+    throw new SmtpClientCommunicationException(
+        'The server does not support LOGIN authentication method.');
+  }
+
+  var username = c.server.username;
+  var password = c.server.password;
+
+  // 'Username:' in base64 is: VXN...
+  await c.send('AUTH LOGIN',
+      acceptedRespCodes: ['334'], expect: 'VXNlcm5hbWU6');
+  // 'Password:' in base64 is: UGF...
+  await c.send(convert.base64.encode(username.codeUnits),
+      acceptedRespCodes: ['334'], expect: 'UGFzc3dvcmQ6');
+  var loginResp = await c
+      .send(convert.base64.encode(password.codeUnits), acceptedRespCodes: []);
+  return loginResp.responseCode.startsWith('2');
+}
+
+Future<bool> _doAuthXoauth2(Connection c) async {
+  var capabilities = c.capabilities;
+  if (!capabilities.authXoauth2) {
+    throw new SmtpClientCommunicationException(
+        'The server does not support XOAUTH2 authentication method.');
+  }
+
+  var token = c.server.xoauth2Token;
+
+  // See https://developers.google.com/gmail/imap/xoauth2-protocol
+  var loginResp = await c.send('AUTH XOAUTH2 $token', acceptedRespCodes: []);
+  return loginResp.responseCode.startsWith('2');
+}
+
+Future<void> _doAuthentication(Connection c) async {
+  bool loginOk = true;
+
+  if (c.server.username != null) {
+    loginOk = await _doAuthLogin(c);
+  } else if (c.server.xoauth2Token != null) {
+    loginOk = await _doAuthXoauth2(c);
+  }
+
+  if (!loginOk) {
+    throw new SmtpClientAuthenticationException(
+        'Incorrect username / password / credentials');
+  }
+}
+
+Future<Connection> connect(SmtpServer smtpServer, Duration timeout) async {
+  final Connection c = new Connection(smtpServer, timeout: timeout);
+
+  try {
+    await c.connect();
+
     try {
-      // Constructor might throw IRProblemException.
-      irMessage = new IRMessage(message);
-    } on IRProblemException catch (e) {
-      sendReports
-          .add(new SendReport(message, false, validationProblems: [e.problem]));
-      return sendReports;
-    }
-
-    try {
-      await _c.connect();
-
       // Greeting (Don't send anything.  We first wait for a 2xx message.)
-      await _c.send(null);
-
-      // EHLO / HELO
-      var capabilities = await _doEhloHelo();
-
-      _c.verifySecuredConnection();
-
-      // Authenticate
-      await _doAuthentication(capabilities);
-
-      Iterable<String> envelopeTos = irMessage.envelopeTos;
-
-      if (envelopeTos.isEmpty) {
-        _logger.info('Mail without recipients.  Not sending. ($message)');
-        sendReports.add(new SendReport(message, false, validationProblems: [
-          new Problem('NO_RECIPIENTS', 'Mail does not have any recipients.')
-        ]));
-        return sendReports;
+      await c.send(null);
+    } on TimeoutException {
+      if (!c.isSecure) {
+        throw new SmtpNoGreetingException(
+            'Timed out while waiting for greeting (try ssl).');
+      } else {
+        throw new SmtpNoGreetingException(
+            'Timed out while waiting for greeting.');
       }
-
-      // Make sure that the server knows, that we are sending a new mail.
-      // This also allows us to simply `continue;` to the next mail in case
-      // something goes wrong.
-      // await _c.send('RSET');  // We currently reconnect for every msg.
-
-      // Tell the server the envelope from address (might be different to the
-      // 'From: ' header!)
-
-      bool smtputf8 = capabilities.smtpUtf8;
-      await _c.send(
-          'MAIL FROM:<${irMessage.envelopeFrom}> ${smtputf8 ? ' SMTPUTF8' : ''}');
-
-      // Give the server all recipients.
-      // TODO what if only one address fails?
-      await Future.forEach(
-          envelopeTos, (recipient) => _c.send('RCPT TO:<$recipient>'));
-
-      // Finally send the actual mail.
-      await _c.send('DATA', acceptedRespCodes: ['2', '3']);
-
-      await _c.sendStream(irMessage.data(capabilities));
-
-      await _c.send('.', acceptedRespCodes: ['2', '3']);
-
-      await _c.send('QUIT', waitForResponse: false);
-
-      sendReports.add(new SendReport(message, true));
-    } catch (exception) {
-      sendReports.add(new SendReport(message, false, validationProblems: [
-        new Problem('UNKNOWN', 'Received an exception: $exception')
-      ]));
-    } finally {
-      await _c.close();
     }
 
-    return sendReports;
+    // EHLO / HELO
+    await _doEhloHelo(c);
+
+    c.verifySecuredConnection();
+
+    // Authenticate
+    await _doAuthentication(c);
+    return c;
+  } catch (e) {
+    if (c != null) {
+      await c.close();
+    }
+    rethrow;
   }
+}
+
+Future<void> close(Connection connection) async {
+  if (connection == null) {
+    return;
+  }
+  try {
+    await connection.send('QUIT', waitForResponse: false);
+  } finally {
+    await connection.close();
+  }
+}
+
+/// Connection [c] must have been opened before.
+/// The message should be validated before passing it to this function.
+/// This function does not close the connection [c].
+///
+/// Throws following exceptions:
+/// [SmtpClientAuthenticationException],
+/// [SmtpClientCommunicationException],
+/// [SmtpUnsecureException],
+/// [SocketException],
+Future<void> sendSingleMessage(
+    Message message, Connection c, Duration timeout) async {
+  IRMessage irMessage = IRMessage(message);
+  Iterable<String> envelopeTos = irMessage.envelopeTos;
+
+  var capabilities = c.capabilities;
+
+  // Tell the server the envelope from address (might be different to the
+  // 'From: ' header!)
+  bool smtpUtf8 = capabilities.smtpUtf8;
+  await c.send(
+      'MAIL FROM:<${irMessage.envelopeFrom}>' + (smtpUtf8 ? ' SMTPUTF8' : ''));
+
+  // Give the server all recipients.
+  // TODO what if only one address fails?
+  await Future.forEach(
+      envelopeTos, (recipient) => c.send('RCPT TO:<$recipient>'));
+
+  // Finally send the actual mail.
+  await c.send('DATA', acceptedRespCodes: ['2', '3']);
+
+  await c.sendStream(irMessage.data(capabilities));
+
+  await c.send('.', acceptedRespCodes: ['2', '3']);
 }